Extract host & manifest to test
